backend/requirements.txt
# === Backend Core ===
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2

# === LangChain & RAG Stack ===
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langchain-huggingface==0.2.0
chromadb==0.4.15
sentence-transformers==2.2.2

# === LLaMA Support ===
llama-cpp-python==0.2.11

# === PDF & NLP ===
PyPDF2==3.0.1
spacy==3.7.2
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0

# === Optional Utilities ===
tqdm>=4.64.0  # progress bars (nice for embedding/debugging)
